Refactor comment form draft lifecycle (#1650)
* refactor(comments): extract form draft lifecycle

* refactor(comments): drop the draft passthrough facade

The extraction left form_controller holding a full mirror of the draft
state: 30 getter/setter pairs and 20 one-line delegates, most of which
nothing called. That kept the old surface alive next to the new owner, so
the controller still read as if it owned draft state.

Point the controller's remaining draft work at the manager directly and
delete every accessor plus the eight delegates with no callers. The
white-box assertions in the existing draft tests follow the state to its
new owner. FormDraftManager's dead `editingId` setter goes too - the
manager only ever reads it.

Adds form_draft_manager.test.js, which drives the manager through a plain
host object instead of Stimulus. That is the point of the extraction: if
the manager reaches for form internals the host does not expose, the test
fails instead of the coupling going unnoticed.
